pyproject.toml
[tool.poetry]
name = "recommender-app"
version = "0.1.0"
description = ""
authors = ["Tommaso Brigaglia <91462387+toretto2002@users.noreply.github.com>"]
readme = "README.md"
packages = [{ include = "recommender_app" }]

[tool.poetry.dependencies]
python = "^3.12"
flask = {version = ">=3.1.1,<4.0.0", extras = ["async"]}
flask-sqlalchemy = ">=3.1.1,<4.0.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
flask-migrate = "^4.1.0"
flask-jwt-extended = "^4.7.1"
flask-cors = "^5.0.0"
pydantic = "^2.11.7"
email-validator = "^2.2.0"
requests = "^2.32.4"
beautifulsoup4 = "^4.13.4"
lxml = "^6.0.0"
playwright = "^1.53.0"
llama-index-llms-ollama = "^0.6.2"
sqlalchemy = "^2.0.41"
llama-index-embeddings-ollama = "^0.6.0"
openai = "^2.6.0"
openai-agents = "^0.4.1"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
